Preferences is now exported from package preferences. updated update_interval to int (was uint16)

+++ internal/preferences/types.go
@@ -4,7 +4,7 @@ type PreferencesFormat struct {
 	TwitchConfig           TwitchConfigT    `json:"twitch_config"`
 	StreamVariables        StreamVariablesT `json:"stream_variables"`
 	LlmVariables           []LlmVariableT   `json:"llm_variables"`
-	VariableUpdateInterval uint16           `json:"variable_update_interval"`
+	VariableUpdateInterval int              `json:"variable_update_interval"`
 	ActivityConsoleOutput  string           `json:"activity_console_output"`
 }
